API de new finalReport agora e closed
