+++ src/route/closeApis/final-report.js
@@ -1,12 +1,9 @@
 const express = require('express');
 const router = express.Router();
 const {
-  addFinalReportTemplate,
-  //  getFinalReportTemplateByVersion,
   getFinalReportData,
 } = require('../../service/final-report');
 const { verifyToken } = require('../../service/authentication');
-const { clone, responseError } = require('../../service/helper');
 
 
 router.get('/by-id', verifyToken, async(req, res) => {
@@ -27,18 +24,6 @@ router.get('/by-id', verifyToken, async(req, res) => {
     console.log(error);
     // TODO error
     res.status(error.code).send(error.message);
-  }
-});
-
-router.post('/new', verifyToken, async(req, res) => {
-  try {
-    const data = clone(req.body);
-    const fr = await addFinalReportTemplate(data);
-
-    res.send(fr._id);
-  } catch (error) {
-    console.log(error);
-    responseError(res, error);
   }
 });
